+++ inst/torch_for_r/torch_PLN.R
@@ -135,3 +135,26 @@ myPLN_torch$loglik
 myPLN_nlopt$loglik
 
 
+data("barents")
+system.time(myPLN_torch <-
+              PLN(Abundance ~ 1 + Depth + Temperature + offset(log(Offset)),
+                  data = barents, control = list(backend = "torch"))
+)
+system.time(myPLN_nlopt <-
+              PLN(Abundance ~ 1 + Depth + Temperature + offset(log(Offset)),
+                  data = barents, control = list(backend = "nlopt", xtol_rel = 1e-8))
+)
+
+par(mfrow = c(2,2))
+plot(myPLN_torch$model_par$Theta,
+     myPLN_nlopt$model_par$Theta); abline(0, 1)
+plot(myPLN_torch$model_par$Sigma,
+     myPLN_nlopt$model_par$Sigma); abline(0, 1)
+plot(myPLN_torch$var_par$M,
+     myPLN_nlopt$var_par$M); abline(0, 1)
+plot(myPLN_torch$var_par$S2,
+     myPLN_nlopt$var_par$S2); abline(0, 1)
+myPLN_torch$loglik
+myPLN_nlopt$loglik
+
+